reorganização do front
