+++ src/main/ui/public/logo512.png
