+++ src/main/ui/public/manifest.json
