+++ src/main/ui/src/service/areas.js
