+++ src/main/ui/src/service/vaga.js
