+++ usersapi/src/main/ui/src/interno/adicionar_remover/RemoverZona.jsx
@@ -1,9 +1,9 @@
 import "./editar_vaga/adicionar_remover.css";
 import { useEffect, useMemo, useState } from "react";
-import VagaService from "../../service/vaga";
+import VagaService from "../../../../../../../src/main/ui/src/service/vaga";
 import Pagination from "../../components/Paginacao/Pagination";
 import { Link } from "react-router-dom";
-import AreasService from "../../service/areas";
+import AreasService from "../../../../../../../src/main/ui/src/service/areas";
 import { Alert, Collapse } from "@mui/material";
 
 let PageSize = 6;
